Silence -Wsign-compare in the H264-in-AVI duration heuristic
VP.Duration (int64_t) was compared directly against vector::size() (size_t).
Cast the size to int64_t; the frame count always fits and the comparison stays
correct, removing the only warnings emitted by the default build.

+++ src/videosource.cpp
@@ -1141,11 +1141,11 @@ bool BestVideoSource::IndexTrack(const ProgressFunction &Progress) {
 
     if (!HasValidPTS) {
         // Probably H264 in AVI
-        if (VP.Duration == TrackIndex.Frames.size()) {
+        if (VP.Duration == static_cast<int64_t>(TrackIndex.Frames.size())) {
             // It's CFR so we know every frame duration is 1 unit
             for (size_t i = 0; i < TrackIndex.Frames.size(); i++)
                 TrackIndex.Frames[i].PTS = i;
-        } else if (VP.Duration == TrackIndex.Frames.size() * 2) {
+        } else if (VP.Duration == static_cast<int64_t>(TrackIndex.Frames.size()) * 2) {
             // It's with 99.99% certainty CFR so we almost know every frame duration is 2 units
             for (size_t i = 0; i < TrackIndex.Frames.size(); i++)
                 TrackIndex.Frames[i].PTS = i * 2;
